+++ src/Config/MessagesConfig.cs
@@ -45,6 +45,11 @@ namespace DiscordConnector.Config
         public MessagesConfig(ConfigFile configFile)
         {
             config = configFile;
+            ReloadConfig();
+        }
+
+        public void ReloadConfig()
+        {
             LoadConfig();
         }
 
@@ -141,11 +146,11 @@ namespace DiscordConnector.Config
             // Messages.Events
             eventStartMessage = config.Bind<string>(EVENT_MESSAGES,
                 "Event Start Message",
-                "**Event**: %EVENT_MSG% around %PLAYERS%",
+                "**Event**: %EVENT_MSG%",
                 "Set the message that will be sent when a random event starts on the server. Sending the coordinates is enabled by default in the toggles config." + Environment.NewLine +
                 "If you want to have this choose from a variety of messages at random, separate each message with a semicolon ';'" + Environment.NewLine +
-                "The special string %EVENT_MSG% will be replaced with the message that is displayed on the screen when the event starts." + Environment.NewLine +
-                "The special string %PLAYERS% will be replaced with a list of players in the event area.");
+                "The special string %EVENT_MSG% will be replaced with the message that is displayed on the screen when the event starts."); // + Environment.NewLine +
+                                                                                                                                            // "The special string %PLAYERS% will be replaced with a list of players in the event area."); //! Removed due to unreliability
             eventStopMessage = config.Bind<string>(EVENT_MESSAGES,
                 "Event Stop Message",
                 "**Event**: %EVENT_MSG%",
@@ -154,19 +159,19 @@ namespace DiscordConnector.Config
                 "The special string %EVENT_MSG% will be replaced with the message that is displayed on the screen when the event stops.");
             eventPausedMessage = config.Bind<string>(EVENT_MESSAGES,
                 "Event Paused Message",
-                "**Event**: %EVENT_END_MSG% -- for now! (Currently paused due to no players in the event area.)",
+                "**Event**: %EVENT_END_MSG% — for now! (Currently paused due to no players in the event area.)",
                 "Set the message that will be sent when a random event is paused due to players leaving the area. Sending the coordinates is enabled by default in the toggles config." + Environment.NewLine +
                 "If you want to have this choose from a variety of messages at random, separate each message with a semicolon ';'" + Environment.NewLine +
                 "The special string %EVENT_START_MSG% will be replaced with the message that is displayed on the screen when the event starts." + Environment.NewLine +
                 "The special string %EVENT_END_MSG% will be replaced with the message that is displayed on the screen when the event ends.");
             eventResumedMessage = config.Bind<string>(EVENT_MESSAGES,
                 "Event Resumed Message",
-                "**Event**: %EVENT_START_MSG% around %PLAYERS%",
+                "**Event**: %EVENT_START_MSG%",
                 "Set the message that will be sent when a random event is resumed due to players re-entering the area. Sending the coordinates is enabled by default in the toggles config." + Environment.NewLine +
                 "If you want to have this choose from a variety of messages at random, separate each message with a semicolon ';'" + Environment.NewLine +
                 "The special string %EVENT_START_MSG% will be replaced with the message that is displayed on the screen when the event starts." + Environment.NewLine +
-                "The special string %EVENT_END_MSG% will be replaced with the message that is displayed on the screen when the event ends." + Environment.NewLine +
-                "The special string %PLAYERS% will be replaced with a list of players in the event area.");
+                "The special string %EVENT_END_MSG% will be replaced with the message that is displayed on the screen when the event ends."); // + Environment.NewLine +
+                                                                                                                                              // "The special string %PLAYERS% will be replaced with a list of players in the event area."); //! Removed due to unreliability
 
             config.Save();
         }