+++ src/Patches/ZNetPatches.cs
@@ -9,22 +9,18 @@ namespace DiscordConnector.Patches
         [HarmonyPatch(typeof(ZNet), nameof(ZNet.LoadWorld))]
         internal class LoadWorld
         {
-            private static void Postfix(ref ZNet __instance)
+            private static void Postfix()
             {
                 if (Plugin.StaticConfig.LoadedMessageEnabled)
                 {
                     DiscordApi.SendMessage(
-                        Plugin.StaticConfig.LoadedMessage.Replace("%PUBLICIP%", Plugin.PublicIpAddress)
+                        MessageTransformer.FormatServerMessage(Plugin.StaticConfig.LoadedMessage)
                     );
                 }
-            }
-            private static void Prefix(ref ZNet __instance)
-            {
-                if (Plugin.StaticConfig.LaunchMessageEnabled)
+
+                if (Plugin.IsHeadless())
                 {
-                    DiscordApi.SendMessage(
-                        Plugin.StaticConfig.LaunchMessage.Replace("%PUBLICIP%", Plugin.PublicIpAddress)
-                    );
+                    Plugin.StaticEventWatcher.Activate();
                 }
             }
         }
@@ -37,31 +33,8 @@ namespace DiscordConnector.Patches
                 if (Plugin.StaticConfig.WorldSaveMessageEnabled)
                 {
                     DiscordApi.SendMessage(
-                        Plugin.StaticConfig.SaveMessage.Replace("%PUBLICIP%", Plugin.PublicIpAddress)
+                        MessageTransformer.FormatServerMessage(Plugin.StaticConfig.SaveMessage)
                     );
-                }
-            }
-        }
-
-        [HarmonyPatch(typeof(ZNet), nameof(ZNet.Shutdown))]
-        internal class Shutdown
-        {
-            private static void Prefix(ref ZNet __instance)
-            {
-                if (Plugin.StaticConfig.StopMessageEnabled)
-                {
-                    DiscordApi.SendMessage(
-                        Plugin.StaticConfig.StopMessage.Replace("%PUBLICIP%", Plugin.PublicIpAddress)
-                        );
-                }
-            }
-            private static void Postfix(ref ZNet __instance)
-            {
-                if (Plugin.StaticConfig.ShutdownMessageEnabled)
-                {
-                    DiscordApi.SendMessage(
-                        Plugin.StaticConfig.ShutdownMessage.Replace("%PUBLICIP%", Plugin.PublicIpAddress)
-                        );
                 }
             }
         }
@@ -88,13 +61,18 @@ namespace DiscordConnector.Patches
                     }
                     if (Plugin.StaticConfig.PlayerDeathMessageEnabled)
                     {
-                        string message = Plugin.StaticConfig.DeathMessage.Replace("%PLAYER_NAME%", peer.m_playerName);
+                        string message = MessageTransformer.FormatPlayerMessage(Plugin.StaticConfig.DeathMessage, peer.m_playerName);
                         if (Plugin.StaticConfig.PlayerDeathPosEnabled)
                         {
-                            DiscordApi.SendMessage(
-                            message,
-                            peer.m_refPos
-                            );
+                            if (Plugin.StaticConfig.DiscordEmbedsEnabled || !message.Contains("%POS%"))
+                            {
+                                DiscordApi.SendMessage(
+                                    message,
+                                    peer.m_refPos
+                                );
+                            }
+                            message = MessageTransformer.FormatPlayerMessage(Plugin.StaticConfig.DeathMessage, peer.m_playerName, peer.m_refPos);
+                            DiscordApi.SendMessage(message);
                         }
                         else
                         {
@@ -103,7 +81,9 @@ namespace DiscordConnector.Patches
                     }
                     if (Plugin.StaticConfig.AnnouncePlayerFirstDeathEnabled && Plugin.StaticRecords.Retrieve(RecordCategories.Death, peer.m_playerName) == 0)
                     {
-                        DiscordApi.SendMessage(Plugin.StaticConfig.PlayerFirstDeathMessage.Replace("%PLAYER_NAME%", peer.m_playerName));
+                        DiscordApi.SendMessage(
+                            MessageTransformer.FormatPlayerMessage(Plugin.StaticConfig.PlayerFirstDeathMessage, peer.m_playerName)
+                        );
                     }
                     if (Plugin.StaticConfig.StatsDeathEnabled)
                     {
@@ -117,13 +97,18 @@ namespace DiscordConnector.Patches
                     Plugin.StaticLogger.LogDebug($"Added player {peer.m_uid} ({peer.m_playerName}) to joined player list.");
                     if (Plugin.StaticConfig.PlayerJoinMessageEnabled)
                     {
-                        string message = Plugin.StaticConfig.JoinMessage.Replace("%PLAYER_NAME%", peer.m_playerName);
+                        string message = MessageTransformer.FormatPlayerMessage(Plugin.StaticConfig.JoinMessage, peer.m_playerName);
                         if (Plugin.StaticConfig.PlayerJoinPosEnabled)
                         {
-                            DiscordApi.SendMessage(
-                            message,
-                            peer.m_refPos
-                            );
+                            if (Plugin.StaticConfig.DiscordEmbedsEnabled || !message.Contains("%POS%"))
+                            {
+                                DiscordApi.SendMessage(
+                                    message,
+                                    peer.m_refPos
+                                );
+                            }
+                            message = MessageTransformer.FormatPlayerMessage(Plugin.StaticConfig.JoinMessage, peer.m_playerName, peer.m_refPos);
+                            DiscordApi.SendMessage(message);
                         }
                         else
                         {
@@ -132,7 +117,9 @@ namespace DiscordConnector.Patches
                     }
                     if (Plugin.StaticConfig.AnnouncePlayerFirstJoinEnabled && Plugin.StaticRecords.Retrieve(RecordCategories.Join, peer.m_playerName) == 0)
                     {
-                        DiscordApi.SendMessage(Plugin.StaticConfig.PlayerFirstJoinMessage.Replace("%PLAYER_NAME%", peer.m_playerName));
+                        DiscordApi.SendMessage(
+                            MessageTransformer.FormatPlayerMessage(Plugin.StaticConfig.PlayerFirstJoinMessage, peer.m_playerName)
+                        );
                     }
                     if (Plugin.StaticConfig.StatsJoinEnabled)
                     {
@@ -152,24 +139,29 @@ namespace DiscordConnector.Patches
                 {
                     if (Plugin.StaticConfig.PlayerLeaveMessageEnabled)
                     {
-                        string message = Plugin.StaticConfig.LeaveMessage.Replace("%PLAYER_NAME%", peer.m_playerName);
+                        string message = MessageTransformer.FormatPlayerMessage(Plugin.StaticConfig.LeaveMessage, peer.m_playerName);
                         if (Plugin.StaticConfig.PlayerLeavePosEnabled)
                         {
-                            DiscordApi.SendMessage(
-                            message,
-                            peer.m_refPos
-                            );
+                            if (Plugin.StaticConfig.DiscordEmbedsEnabled || !message.Contains("%POS%"))
+                            {
+                                DiscordApi.SendMessage(
+                                    message,
+                                    peer.m_refPos
+                                );
+                            }
+                            message = MessageTransformer.FormatPlayerMessage(Plugin.StaticConfig.LeaveMessage, peer.m_playerName, peer.m_refPos);
+                            DiscordApi.SendMessage(message);
                         }
                         else
                         {
-                            DiscordApi.SendMessage(
-                            message
-                            );
+                            DiscordApi.SendMessage(message);
                         }
                     }
                     if (Plugin.StaticConfig.AnnouncePlayerFirstLeaveEnabled && Plugin.StaticRecords.Retrieve(RecordCategories.Leave, peer.m_playerName) == 0)
                     {
-                        DiscordApi.SendMessage(Plugin.StaticConfig.PlayerFirstLeaveMessage.Replace("%PLAYER_NAME%", peer.m_playerName));
+                        DiscordApi.SendMessage(
+                            MessageTransformer.FormatPlayerMessage(Plugin.StaticConfig.PlayerFirstLeaveMessage, peer.m_playerName)
+                        );
                     }
                     if (Plugin.StaticConfig.StatsLeaveEnabled)
                     {