+++ src/Plugin.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Timers;
-using BepInEx;
+﻿using BepInEx;
 using BepInEx.Logging;
 using HarmonyLib;
 using UnityEngine;
@@ -15,6 +12,8 @@ namespace DiscordConnector
         internal static ManualLogSource StaticLogger;
         internal static PluginConfig StaticConfig;
         internal static Records StaticRecords;
+        internal static Leaderboard StaticLeaderboards;
+        internal static EventWatcher StaticEventWatcher;
         internal static string PublicIpAddress;
 #if !NoBotSupport
         private static Webhook.Listener _listener;
@@ -26,6 +25,7 @@ namespace DiscordConnector
             StaticLogger = Logger;
             StaticConfig = new PluginConfig(Config);
             StaticRecords = new Records(Paths.GameRootPath);
+            StaticLeaderboards = new Leaderboard();
         }
 
         private void Awake()
@@ -36,6 +36,10 @@ namespace DiscordConnector
             {
                 StaticLogger.LogInfo("Not running on a dedicated server, some features may break -- please report them!");
             }
+            else
+            {
+                StaticEventWatcher = new EventWatcher();
+            }
 
             if (string.IsNullOrEmpty(StaticConfig.WebHookURL))
             {
@@ -45,7 +49,9 @@ namespace DiscordConnector
             if (StaticConfig.StatsAnnouncementEnabled)
             {
                 System.Timers.Timer leaderboardTimer = new System.Timers.Timer();
-                leaderboardTimer.Elapsed += SendLeaderboardAnnouncement;
+                leaderboardTimer.Elapsed += StaticLeaderboards.OverallHighest.SendLeaderboardOnTimer;
+                leaderboardTimer.Elapsed += StaticLeaderboards.OverallLowest.SendLeaderboardOnTimer;
+                leaderboardTimer.Elapsed += StaticLeaderboards.TopPlayers.SendLeaderboardOnTimer;
                 // Interval is learned from config file in minutes
                 leaderboardTimer.Interval = 60 * 1000 * StaticConfig.StatsAnnouncementPeriod;
                 leaderboardTimer.Start();
@@ -74,48 +80,6 @@ namespace DiscordConnector
                 _listener.Dispose();
             }
 #endif
-        }
-
-        /// <summary>
-        /// Function which is a valid Timer action to provide sending leader board messages on an interval.
-        /// 
-        /// This should be moved to its own class and we should provide separate private functions to handle each
-        /// variety of leader board. This enables greater flexibility when it comes to what kinds of leader boards
-        /// can be supported.
-        /// </summary>
-        private void SendLeaderboardAnnouncement(object sender, ElapsedEventArgs elapsedEventArgs)
-        {
-            var deathLeader = StaticRecords.Retrieve(RecordCategories.Death);
-            var joinLeader = StaticRecords.Retrieve(RecordCategories.Join);
-            var shoutLeader = StaticRecords.Retrieve(RecordCategories.Shout);
-            var pingLeader = StaticRecords.Retrieve(RecordCategories.Ping);
-
-            List<Tuple<string, string>> leaderFields = new List<Tuple<string, string>>();
-            if (StaticConfig.LeaderboardDeathEnabled && deathLeader.Item2 > 0)
-            {
-                leaderFields.Add(Tuple.Create("Most Deaths", $"{deathLeader.Item1} ({deathLeader.Item2})"));
-            }
-            if (StaticConfig.LeaderboardSessionEnabled && joinLeader.Item2 > 0)
-            {
-                leaderFields.Add(Tuple.Create("Most Sessions", $"{joinLeader.Item1} ({joinLeader.Item2})"));
-            }
-            if (StaticConfig.LeaderboardShoutEnabled && shoutLeader.Item2 > 0)
-            {
-                leaderFields.Add(Tuple.Create("Most Shouts", $"{shoutLeader.Item1} ({shoutLeader.Item2})"));
-            }
-            if (StaticConfig.LeaderboardPingEnabled && pingLeader.Item2 > 0)
-            {
-                leaderFields.Add(Tuple.Create("Most Pings", $"{pingLeader.Item1} ({pingLeader.Item2})"));
-            }
-            if (leaderFields.Count > 0)
-            {
-                DiscordApi.SendMessageWithFields("Current leader board for tracked stats:", leaderFields);
-            }
-            else
-            {
-                StaticLogger.LogInfo("Not sending a leaderboard because theirs either no leaders, or nothing allowed.");
-            }
-
         }
 
         /// <summary>